finally got edit/upload working correctly
