+++ src/App.jsx
@@ -3,7 +3,6 @@ import { HashRouter, Routes, Route, Link } from "react-router-dom";
 import InitialScreen from "./Screens/InitialScreen.jsx";
 import ImageUploadScreen from "./Screens/ImageUploadScreen.jsx";
 import NoteUploadScreen from "./Screens/NoteUploadScreen.jsx";
-import EditNoteScreen from "./Screens/EditNoteScreen.jsx";
 export default function App() {
   return (
     <HashRouter>
@@ -14,7 +13,6 @@ export default function App() {
         <Route path="/" element={<InitialScreen />} />
         <Route path="/imageuploadscreen" element={<ImageUploadScreen />} />
         <Route path="/noteuploadscreen" element={<NoteUploadScreen />} />
-        <Route path="/editnotescreen" element={<EditNoteScreen />} />
       </Routes>
     </HashRouter>
   );
