+++ src/Screens/InitialScreen.jsx
@@ -1,7 +1,6 @@
 import React from "react";
 import NoteUploader from "../Components/NoteUploader.jsx";
 import SexyImage from "../Components/SexyImage.jsx";
-import MapRepoFolder from "../Components/MapRepoFolder.jsx";
 import Products from "../Components/Products.jsx";
 import { useNavigate } from "react-router-dom";
 const InitialScreen = () => {
@@ -14,10 +13,9 @@ const InitialScreen = () => {
       <button onClick={() => navigate("/imageuploadscreen")}>
         Upload Image
       </button>
-      <button onClick={() => navigate("/editnotescreen")}>Edit Notes</button>
+
       {/* <Products /> */}
       <NoteUploader />
-      <MapRepoFolder />
       <SexyImage />
     </div>
   );
